+++ src/test/java/com/imsweb/datagenerator/naaccr/rule/patient/NameRuleTest.java
@@ -10,7 +10,7 @@ import org.junit.Test;
 
 public class NameRuleTest {
 
-    private NameRule _rule = new NameRule();
+    private final NameRule _rule = new NameRule(false);
 
     @Test
     public void testExecute() {
@@ -28,14 +28,14 @@ public class NameRuleTest {
                     Assert.assertTrue(rec.get("nameFirst").matches("^[a-zA-Z]+$"));
                     Assert.assertTrue(rec.get("nameMiddle").matches("^[a-zA-Z]+$") || rec.get("nameMiddle").matches("^[A-Z]\\.$"));
                     Assert.assertTrue(rec.get("nameLast").matches("^[a-zA-Z]+$"));
-                    Assert.assertTrue(rec.get("namePrefix").matches("^[\\w \\.]*$"));
-                    Assert.assertTrue(rec.get("nameSuffix").matches("^[\\w \\.]*$"));
+                    Assert.assertTrue(rec.get("namePrefix").matches("^[\\w .]*$"));
+                    Assert.assertTrue(rec.get("nameSuffix").matches("^[\\w .]*$"));
 
                     // verify male patient was not given a maiden name, else verify female maiden name if present
                     if (sex.equals("1"))
-                        Assert.assertTrue(rec.get("nameMaiden").isEmpty());
+                        Assert.assertTrue(rec.get("nameBirthSurname").isEmpty());
                     else
-                        Assert.assertTrue(rec.get("nameMaiden").isEmpty() || rec.get("nameMaiden").matches("^[a-zA-Z]+$"));
+                        Assert.assertTrue(rec.get("nameBirthSurname").isEmpty() || rec.get("nameBirthSurname").matches("^[a-zA-Z]+$"));
                 }
             }
         }