+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Synthetic data generator
 
+**Changes in version 1.18**
+
+- Fixed NAACCR XML generator not providing the user-defined dictonaries to the XML writer.
+- Changed name rule so it assigns field 'nameBirthSurname' instead of 'nameMaiden' for NAACCR version 210 and later.
+- Updated Layout library from version 2.0 to version 3.1.
+- Updated Algorithms library from version 2.14 to version 2.18.
+- Updated Staging client library from version 4.8 to version 5.0.
+- Updated Commons Lang library from version 3.10 to version 3.11.
+
 **Changes in version 1.17**
 
 - Fixed NAACCR XML generator so it returns Patient objects and not list of records.